requirements.txt
google-generativeai
python-telegram-bot
pyautogui
pillow
python-dotenv
